requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
async-timeout==4.0.2
asyncio==3.4.3
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.2.0
cod-api==2.0.1
DateTime==5.2
discord.py==2.3.1
emoji==2.7.0
enum34==1.1.10
exceptiongroup==1.1.2
frozenlist==1.4.0
h11==0.14.0
idna==3.4
multidict==6.0.4
outcome==1.2.0
PySocks==1.7.1
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
selenium==4.11.2
sniffio==1.3.0
sortedcontainers==2.4.0
trio==0.22.2
trio-websocket==0.10.3
urllib3==2.0.4
uuid==1.30
wsproto==1.2.0
yarl==1.9.2
zope.interface==6.0
